Made Episode and Season callable.  __call__ emulates behavior of Show.__call__
Although this functionality can be found in Episode.get and Season.get, it is
useful because it provides an interface consitent with that of Series.

This is particularly useful in that it allows for simple automated obtention of
values, for example with `map`:

map(lambda obj, arg: obj(arg), [obj1, obj2, obj3], [arg1, arg2, arg3])

In the above example, objs can be Shows Seasons or Episodes.