Adding option for DVD ordering
This corresponds with a PR I am making for the Sick-Beard project that allows for ordering the show episodes by their DVD order instead of their airdate order.

https://github.com/midgetspy/Sick-Beard/pull/742

This is an important option for shows like "Batman: The Animated Series" and "Firefly" where the network screwed up the air ordering but fixed it in a subsequent DVD release.

Adding Tests